requirements.txt
fastapi
uvicorn
motor
python-dotenv
pandas
numpy
matplotlib
seaborn
scikit-learn
scipy
pydantic